requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
Werkzeug==3.1.3
blinker==1.9.0
boto3==1.35.68
botocore==1.35.68
click==8.1.7
Flask==3.1.0
Flask-Mail==0.10.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.3
